feat: mvp

+++ app/Http/Controllers/ZoneBulkDeleteController.php
@@ -11,6 +11,8 @@ class ZoneBulkDeleteController extends Controller
 {
     public function __invoke(): RedirectResponse
     {
+        abort_if(!auth()->user()->admin, 403);
+
         Zone::truncate();
 
         Cache::forget('zones');
